Ospats Function from Ospats-package

Copy-paste of the original code (with small edits)

Usage

ospatsF(
  data = NULL,
  dRange = NULL,
  nCycles = NULL,
  dStart = NULL,
  ClusterStart = NULL,
  dMaxrun = NULL,
  dRSquare = NULL,
  dStrata = NULL,
  initialTemperature = NULL,
  coolingRate = NULL,
  debug = FALSE,
  verbose = TRUE
)

Arguments

data

data matrix with columns x y pred var in that specific order

dRange

Range parameter of the exponential correlation model

nCycles

Number of iterations per one run of the optimisation algorithm

dStart

choose between kMeans (0) or CumrootSquare(1) or external (3)

ClusterStart

external for dStart == 3 (Saby Input)

dMaxrun

Number of independent (apart from same initial states) repeated optimisation runs

dRSquare

R^2 parameter for adjusting for "regression-to-the-mean" (see the paper)

dStrata

Number of strata to have

initialTemperature

simulated annealing parameter

coolingRate

simulated annealing parameter

debug

Useful during development for troubleshooting issues

verbose

Print runtime diagnostics

Details

These are the refactoring changes to the original code to minimize the footprint:

The call fastPdist2(Ar = as.matrix(z), Br = as.matrix(z)))^2 is now outer(z, z, "-")^2

The call fastVsum(Ar = as.matrix(s2)) is now outer(s2, s2, "+")

The call fastPdist2(Ar = xy, Br = xy) is now as.matrix(dist(xy))

Other than that the code has not been touched. Only this documentation has been added.

Original documentation

opstats algorithm:

As described in de Gruijter et al. (2015) Optimizing Stratification and Allocation for Design-Based Estimation of Spatial Means Using Predictions with Error

What does it do:

Given a map (for instance a digital soil map) with prediction variance this function will derive a spatial stratifcation somewhere on the spectrum between and including compact geographical stratification (high uncertainty) and clusters based on the quantiles of the data (low uncertainty).